+++ packages/components/src/form-token-field/index.tsx
@@ -753,7 +753,6 @@ export function FormTokenField( props: FormTokenFieldProps ) {
 				<StyledHelp
 					id={ `components-form-token-suggestions-howto-${ instanceId }` }
 					className="components-form-token-field__help"
-					__nextHasNoMarginBottom
 				>
 					{ tokenizeOnSpace
 						? __(