+++ packages/ui/src/box/stories/index.story.tsx
@@ -1,16 +1,6 @@
-/**
- * External dependencies
- */
 import { type Meta, type StoryObj } from '@storybook/react';
-
-/**
- * WordPress dependencies
- */
+import { type PaddingSize } from '@wordpress/theme';
 import '@wordpress/theme/design-tokens.css';
-
-/**
- * Internal dependencies
- */
 import { Box } from '../box';
 
 const meta: Meta< typeof Box > = {
@@ -35,7 +25,7 @@ export const Default: Story = {
 	argTypes: {
 		padding: {
 			control: 'select',
-			options: [ '2xs', 'xs', 'sm', 'md', 'lg', 1, 2, 3, 4 ],
+			options: [ '2xs', 'xs', 'sm', 'md', 'lg' ] satisfies PaddingSize[],
 		},
 	},
 };